// src/components/lib/auth/types.d.ts
export type User = {
  id: string;
  fullName: string;
  email: string;
  phoneNumber: string;
  dateOfBirth: string;
  gender: string;
  address: string;
  addressOpt: string | null;
  localState: string | null;
  country: string | null;
  customerPath: string;
  status: string;
  dateModified: string;
  dateCreated: string;
};